refactor: Components 분리
